# Todo/docs/swagger.yaml
definitions:
  models.SignInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.TodoItem:
    properties:
      description:
        type: string
      done:
        type: boolean
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.TodoList:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  models.UpdateItemInput:
    properties:
      description:
        type: string
      done:
        type: boolean
      title:
        type: string
    type: object
  models.UpdateListInput:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  models.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /api/items/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting an item by a its id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: item_id
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            type: string
      summary: Delete an Item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get item by specific id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoItem'
      summary: Get item by Id
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Updating an item by a its id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' update item info'
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpdateItemInput'
      - description: item_id
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            type: string
      summary: Update an Item
      tags:
      - items
  /api/lists:
    get:
      consumes:
      - application/json
      description: Get all todo lists of current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: onject
      summary: Get all lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: Creating new todo list  for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: list info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.TodoList'
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: list_id
          schema:
            type: string
      summary: Create List
      tags:
      - lists
  /api/lists/{id}:
    delete:
      consumes:
      - application/json
      description: Deleting a list by a its id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            type: string
      summary: Delete a list
      tags:
      - lists
    get:
      consumes:
      - application/json
      description: Get list by a specific id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TodoItem'
      summary: Get list by Id
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: Updating a List by a its id for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: ' update item info'
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.UpdateListInput'
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            type: string
      summary: Update a List
      tags:
      - lists
  /api/lists/{list_id}/items:
    get:
      consumes:
      - application/json
      description: Get all items from the given list of current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: onject
      summary: Get all items
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Creating new item inside the list for current user
      parameters:
      - description: Bearer Auth, pls add bearer before
        in: header
        name: Authorization
        required: true
        type: string
      - description: item info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.TodoItem'
      - description: list_id
        in: path
        name: list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: id
          schema:
            type: string
      summary: Create Item
      tags:
      - items
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: logging in the user
      parameters:
      - description: username password
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: someToken
          schema:
            type: string
      summary: sign in
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: account info
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      summary: sign up
      tags:
      - auth
swagger: "2.0"